+++ code/globals.h
@@ -254,7 +254,7 @@ extern bool							drag_select_aborted;
 
 extern GroundType  				Ground[LAND_COUNT];
 
-extern int LParam;
+extern intptr_t LParam;
 
 /*
 **	Constant externs (data is not modified during game play).
